requirements.txt
fastapi
uvicorn
alembic
sqlmodel
sqlalchemy>=2.0
psycopg[binary]
python-jose[cryptography]
python-dotenv
email-validator
pydantic[email]
python-multipart
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
